Added "Convert a Variable"
**__INFO (What it does)__**:
Basically, it Converts your input or Variable into a Number or String.
String is normally those with `""`/`''`/ ` `` `
Numbers are as expected Numbers but without those "" things.
These Numbers will be saved if the input was a word as "NaN" (Not available Now) because the Input weren't a Number (Convert: To Number).
Strings are mostly saved as `"Hello"`. Strings aren't affected by anything so the input will be the same as before, but if it were a Number before, it Convert's into a String. Useful to Convert a saved json from webapi which it was a number before.
More Information  for Strings in  Javascript(Programming language.): ( https://www.w3schools.com/jsref/jsref_obj_string.asp ),
More Information for Numbers(Int) in Javascript(Programming lanquage.): ( https://www.w3schools.com/js/js_numbers.asp )

The Mod is under the Section "Other Stuff"